refactor(assets): rename mask to components
