+++ app/gather/assets/apps/survey/Survey.jsx
@@ -35,7 +35,7 @@ import { patchData } from '../utils/request'
 import { FetchUrlsContainer, PaginationContainer } from '../components'
 import SurveyDetail from './SurveyDetail'
 import SurveyDashboard from './SurveyDashboard'
-import SurveyMasks from './mask/SurveyMasks'
+import SurveyMasks from './components/SurveyMasks'
 import EntitiesList from './entity/EntitiesList'
 import EntityItem from './entity/EntityItem'
 import EntitiesDownload from './entity/EntitiesDownload'
